requirements.txt
icalendar>=5.0.11
requests>=2.31.0
python-dotenv>=1.0.0
